+++ UIKitStarterPack/Object_Movie/Object_Movie/Common/Audience.swift
@@ -1,0 +1,20 @@
+//
+//  Audience.swift
+//  Object_Movie
+//
+//  Created by Greem on 4/2/25.
+//
+
+import Foundation
+
+class Audience {
+    
+    
+    func setInvitation(invitation: Invitation) {
+        
+    }
+    
+    func getTicket() -> Ticket {
+        
+    }
+}
