+++ UIKitStarterPack/Object_Movie/Object_Movie/Common/Invitation.swift
@@ -1,0 +1,15 @@
+//
+//  Invitation.swift
+//  Object_Movie
+//
+//  Created by Greem on 4/2/25.
+//
+
+import Foundation
+
+class Invitation {
+    weak var theater: Theater!
+    init(_ theater: Theater) {
+        self.theater = theater
+    }
+}
