+++ UIKitStarterPack/Object_Movie/Object_Movie/Common/Theater.swift
@@ -7,101 +7,7 @@
 
 import Foundation
 
-class Audience {
-    
-    
-    func setInvitation(invitation: Invitation) {
-        
-    }
-    
-    func getTicket() -> Ticket {
-        
-    }
-}
 
-class Ticket: Hashable,Equatable {
-    private var id = UUID()
-    
-    weak var theater: Theater!
-    private var isEntered: Bool = false
-    
-    init(_ theater: Theater){
-        self.theater = theater
-    }
-    
-    func isValid(theater:Theater) -> Bool {
-        if isEntered || self.theater != theater || self.theater == nil{
-            return false
-        } else {
-            isEntered = true
-            return true
-        }
-    }
-    
-    var getFee: Double {
-        self.theater.fee
-    }
-    
-    func hash(into hasher: inout Hasher) {
-        hasher.combine(id)
-    }
-    static func == (lhs: Ticket, rhs: Ticket) -> Bool { lhs.id == rhs.id }
-}
-
-class Invitation {
-    weak var theater: Theater!
-    init(_ theater: Theater) {
-        self.theater = theater
-    }
-}
-
-class TicketOffice:Hashable,Equatable {
-    private let id: UUID = UUID()
-    
-    private var amount: Int = 0
-    private var tickets: [Ticket] = []
-    
-    init(amount: Int) {
-        self.amount = amount
-    }
-    
-    static func == (lhs: TicketOffice, rhs: TicketOffice) -> Bool {
-        lhs.id == rhs.id
-    }
-    func hash(into hasher: inout Hasher) {
-        hasher.combine(id)
-    }
-    
-    func addTicket(ticket: Ticket) {
-        self.tickets.append(ticket)
-    }
-    
-    
-    func getTicketWithFee() -> Ticket? {
-        if(tickets.isEmpty) {
-            return nil
-        }
-        else {
-            let ticket:Ticket =  tickets.removeFirst();
-            self.amount += Int(ticket.getFee)
-            return ticket
-        }
-    }
-    
-    
-    func getTicketWithNoFee() -> Ticket? {
-        if(tickets.isEmpty) {
-            return nil
-        }
-        else {
-            return tickets.removeFirst();
-        }
-    }
-    
-    func getTicketPrice() -> Double {
-        return tickets.first?.getFee ?? 0
-    }
-}
 
 class Theater:Equatable {
     private var id = UUID()
